refactor(mobile): 알림 설정 상세 화면 구현 및 UI 개선
- 날씨/리마인드 알림 상세 설정 화면 추가
- 기존 단일 notifications.tsx 삭제
- NavigationRow 패딩 및 켜짐/꺼짐 컬러 구분 적용
- 날씨 화면 설정 아이콘이 날씨 알림 상세로 직접 이동
