+++ apps/mobile/src/features/notification/presentations/components/settings/NavigationRow.tsx
@@ -1,22 +1,36 @@
-import { ArrowRightIcon, HStack } from '@src/shared/ui';
+import { ArrowRightIcon, HStack, Text } from '@src/shared/ui';
 import { cn } from '@src/shared/utils/cn';
-import { Description, Label, PressableFeedback } from 'heroui-native';
+import { Label, PressableFeedback } from 'heroui-native';
 
 interface NavigationRowProps {
   label: string;
   summary: string;
+  summaryEnabled?: boolean;
   onPress: () => void;
   isDisabled?: boolean;
 }
 
-export function NavigationRow({ label, summary, onPress, isDisabled }: NavigationRowProps) {
+export function NavigationRow({
+  label,
+  summary,
+  summaryEnabled,
+  onPress,
+  isDisabled,
+}: NavigationRowProps) {
   return (
     <PressableFeedback onPress={onPress} isDisabled={isDisabled} className="rounded-lg">
       <PressableFeedback.Highlight className="rounded-lg" />
-      <HStack justify="between" align="center" className={cn(isDisabled && 'opacity-40')} gap={20}>
+      <HStack
+        justify="between"
+        align="center"
+        className={cn('px-4 py-2', isDisabled && 'opacity-40')}
+        gap={20}
+      >
         <Label>{label}</Label>
         <HStack gap={8} align="center">
-          <Description>{summary}</Description>
+          <Text size="b2" className={summaryEnabled ? 'text-main' : 'text-gray-5'}>
+            {summary}
+          </Text>
           <ArrowRightIcon colorClassName="text-gray-6" />
         </HStack>
       </HStack>
